added login required decorator and if statement to check if user is authenticated to make changes

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -69,8 +70,12 @@ def single_product_detail(request, product_id):
 
     return render(request, 'products/single_product_detail.html', context)
 
-
+@login_required
 def add_a_product(request):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to carry out this task')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -90,8 +95,13 @@ def add_a_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_a_product(request, product_id):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to carry out this task')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -113,8 +123,12 @@ def edit_a_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_a_product(request, product_id):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to carry out this task')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
